+++ examples/simple/app.ts
@@ -1,12 +1,12 @@
-import * as crea from '../../dist/index';
+import * as restla from '../../dist/index';
 
 import { db } from './db';
 import { mockPresidents }from './mocks';
 import { presidentsResource } from './resources/president';
 
 let port = process.env.PORT || 9000;
-let app = new crea.Application();
-let router = new crea.Router();
+let app = new restla.Application();
+let router = new restla.Router();
 
 router.use('/presidents', presidentsResource.routes());
 app.use(router.routes());
@@ -17,6 +17,7 @@ db
   .then(async (): Promise<void> => {
     await mockPresidents(db);
 
-    app.listen(port);
-    console.log('Crea running on port: %d', port);
+    app.listen(port, () => {
+      console.log('Listening on port: %d', port);
+    });
   });