+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.2.0
+
+* Correct error handling, and have all error types correctly support the instanceof operator
+* Move to Squell & ModelSafe 0.5
+* Add `associations` flag for including all associations in a resource body
+
 # 0.1.0
 
 * Initial release